refactor: Independent test suites for major features
