+++ machine/lib/vault/vault_managers.py
@@ -28,11 +28,11 @@ Feature managers should not:
 - Depend on each other unless the features explicitly require the dependency.
 """
 
-from dataclasses import dataclass
 import json
 import logging
 import os
 from abc import ABC, abstractmethod
+from dataclasses import dataclass
 from datetime import datetime, timedelta
 from enum import Enum, auto
 from typing import FrozenSet, MutableMapping, TextIO
@@ -104,6 +104,7 @@ path "sys/internal/ui/mounts/{mount}" {{
   capabilities = ["read"]
 }}
 """
+
 
 class LogAdapter(logging.LoggerAdapter):
     """Adapter for the logger to prepend a prefix to all log lines."""
