+++ machine/tox.ini
@@ -17,7 +17,7 @@ all_path = {[vars]src_path} {[vars]unit_test_path} {[vars]integration_test_path}
 runner = uv-venv-lock-runner
 with_dev = true
 set_env =
-    PYTHONPATH = {toxinidir}:{tox_root}/lib:{[vars]src_path}
+    PYTHONPATH = {toxinidir}:{toxinidir}/lib:{[vars]src_path}
     PYTHONBREAKPOINT=pdb.set_trace
     PY_COLORS=1
 deps =
